requirements.txt
openai
yt-dlp
pysrt
python-dotenv